requirements.txt
streamlit~=1.12.0
altair==4.0
requests
oauthlib
python-dotenv
pymongo
PyPDF2